feat: Implement Zod validation in controllers and add Review controller and routes
- Integrate Zod validation in all controllers to ensure data consistency and improve error handling.
- Add a new Review controller with functions for adding, updating, and deleting reviews.
- Include corresponding routes for the Review controller in the application.

+++ backend/controllers/courseCategory.controller.js
@@ -1,5 +1,18 @@
 import express from 'express';
+import { z } from 'zod';
 import CourseCategory from '../models/CourseCategory.model.js';
+import Course from '../models/Course.model.js';
+
+const categorySchema = z.object({
+    categoryName: z.string().min(6).max(50),
+    description: z.string().min(10),
+    accessRestrictions: z.coerce.number()
+                            .int()
+                            .gte(1)
+                            .lte(50),
+    coverImageUrl: z.string()
+})
+const categoryIdSchema = z.string().uuid();
 
 const courseCategoryController = {
     addCategory: async (req, res) => {
@@ -13,6 +26,14 @@ const courseCategoryController = {
                 accessRestrictions,
                 coverImageUrl,
             })
+
+            const validationResult = categorySchema.safeParse(courseCategory);
+            if(!validationResult.success){
+                return res.status(400).json({
+                    error: 'Invalid course categry format',
+                    details: validationResult.error.errors
+                })
+            }
 
             await courseCategory.save();
 
@@ -41,6 +62,13 @@ const courseCategoryController = {
                 accessRestrictions,
                 coverImageUrl,
             })
+            const validationResult = categorySchema.safeParse(updatedCategory);
+            if(!validationResult.success){
+                return res.status(400).json({
+                    error: 'Invalid course categry format',
+                    details: validationResult.error.errors
+                })
+            }
             const courseCategory = await CourseCategory.findByIdAndUpdate(courseCategoryId, updatedCategory);
 
             if(!courseCategory){
@@ -63,6 +91,13 @@ const courseCategoryController = {
     removeCategory: async(req, res) => {
         try{
             const courseCategoryId = req.params.courseCategoryId;
+            const validationResult = categoryIdSchema.safeParse(courseCategoryId);
+            if(!validationResult.success){
+                return res.status(400).json({
+                    error: 'Invalid category id format',
+                    details: validationResult.error.errors
+                })
+            }
 
             const courseCategory = await CourseCategory.findByIdAndDelete(courseCategoryId);
             if(!courseCategory){
@@ -85,7 +120,13 @@ const courseCategoryController = {
     getCourseByCategoryId : async (eq,res) => {
         try{
             const courseCategoryId = req.params.courseCategoryId;
-
+            const validationResult = categoryIdSchema.safeParse(courseCategoryId);
+            if(!validationResult.success){
+                return res.status(400).json({
+                    error: 'Invalid category id format',
+                    details: validationResult.error.errors
+                })
+            }
             const courseCategory = await CourseCategory.findById(courseCategoryId);
             if(!courseCategory){
                 return res.status(404).json({
